Improved returning method and procedural farming Added item blacklist when depositing items Added better sleep timer for Harvest Turtle position starts now at 0,0 Added in farm position which is turtle position offset by one

+++ TurtleTools.lua
@@ -1,10 +1,13 @@
 --TurtleTools 1.0.0 by Raikokiar
+local tableUtils = require("TableUtils")
 
 INVENTORY_SIZE = 16
 
 LOW_FUEL_THRESHOLD = nil
 RepeatUntilRefilled = true
 TryRefillIfLow = true
+
+ItemDropBlacklist = {}
 
 PrintingMethod = print
 
@@ -58,12 +61,19 @@ function IsFull()
 end
 
 function DropInventory()
+    local hasBlacklist = table.maxn(ItemDropBlacklist) > 0
     for i = 1, 16, 1 do
         if turtle.getItemCount(i) > 0 then
             turtle.select(i)
         end
         if not turtle.refuel(0) then
-            turtle.drop()
+            if hasBlacklist and turtle.getItemCount() > 0 then
+                if not TableContainsValue(ItemDropBlacklist, turtle.getItemDetail().name) then
+                    turtle.drop()
+                end
+            else
+                turtle.drop()
+            end
         end
     end
 end
@@ -74,8 +84,20 @@ function DebugLog(...)
     end
 end
 
+--Add item to not be drop when dropping inveotry
+function AddItemToBlacklist(itemID)
+    table.insert(ItemDropBlacklist, itemID)
+end
+
 function SetPrintingMethod(method)
     PrintingMethod = method
 end
 
-return { SelectItem = SelectItem, MoveOrRefuel = MoveOrRefuel, IsFull = IsFull, DropInventory = DropInventory, SetPrintingMethod = SetPrintingMethod }
+return {
+    SelectItem = SelectItem,
+    MoveOrRefuel = MoveOrRefuel,
+    IsFull = IsFull,
+    DropInventory = DropInventory,
+    SetPrintingMethod = SetPrintingMethod,
+    AddItemToBlacklist = AddItemToBlacklist
+}
